third_party/WebKit/LayoutTests/fast/css/cssimagesetvalue-crash.html

<html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</style>
<div id="target"></div>
<script>
target.style.backgroundImage = "-webkit-image-set(url(test1.png) 1x, url(test2.png) 2x)";
test(function() {
  var expected = '-webkit-image-set(url("' + location.href.replace('cssimagesetvalue-crash.html', 'test1.png') + '") 1x, url("' + location.href.replace('cssimagesetvalue-crash.html', 'test2.png') + '") 2x)';
  assert_equals(getComputedStyle(target).backgroundImage, expected);
  assert_equals(target.style.backgroundImage, '-webkit-image-set(url("test1.png") 1x, url("test2.png") 2x)');
}, 'Testing serialization of -webkit-image-set');

</script>
</html>
